refactor structure of io package
