+++ examples/basic/ds1.csv
@@ -1,0 +1,5 @@
+doc_id,doc_value
+1,hello
+2,world
+3,foo
+4,bar
