+++ examples/basic/ds2.jl
@@ -1,0 +1,5 @@
+{"ident": "a", "values":["a1"]}
+{"ident": "b", "values":["b1", "b2"]}
+{"ident": "c", "values":["c1"]}
+
+{"ident": "d"}
