--- examples/ex1/df_corpus_1.txt
@@ -1,3 +1,0 @@
-a b a
-a c
-a
