--- examples/ex1/ex1.py
@@ -1,1 +1,0 @@
-import rltk as tk
