--- examples/ex1/jl_file_1.jsonl
@@ -1,2 +1,0 @@
-{"id":1, "desc":["abc", "def"]}
-{"id":2, "desc":["hij", "def"]}
