--- examples/ex4/blocking.jsonl
@@ -1,4 +1,0 @@
-{"1":["10"]}
-{"2":["11"]}
-{"3":["12", "13"]}
-{"4":["13"]}
