--- examples/ex4/ex_data1.csv
@@ -1,5 +1,0 @@
-id,value
-1,hello
-2,world
-3,foo
-4,bar
