--- examples/ex4/ex_data2.csv
@@ -1,5 +1,0 @@
-id,value
-10,hallo
-11,world
-12,foo
-13,bay
