--- examples/ex4/label.jsonl
@@ -1,10 +1,0 @@
-{"id": ["1","10"], "label": 1.0}
-{"id": ["3","12"], "label": 1.0}
-{"id": ["4","13"], "label": 1.0}
-{"id": ["2","13"], "label": 0.0}
-{"id": ["2","10"], "label": 0.0}
-{"id": ["3","10"], "label": 0.0}
-{"id": ["3","13"], "label": 0.0}
-{"id": ["4","11"], "label": 0.0}
-{"id": ["4","12"], "label": 0.0}
-{"id": ["1","12"], "label": 0.0}
