--- examples/ex4/predicted.jsonl
@@ -1,5 +1,0 @@
-{"predicted_label": 1.0, "id": ["1", "10"], "probability": 0.92157917022495972}
-{"predicted_label": 1.0, "id": ["2", "11"], "probability": 0.96811466410598424}
-{"predicted_label": 1.0, "id": ["3", "12"], "probability": 0.96811466410598424}
-{"predicted_label": 0.0, "id": ["3", "13"], "probability": 0.068006099330006231}
-{"predicted_label": 1.0, "id": ["4", "13"], "probability": 0.82992509358590183}
