--- examples/file_iterator/file_iter_test.csv
@@ -1,8 +1,0 @@
-id,content
-id1,hello
-id2,world
-id3,foo
-id4,"bar
-bar2
-bar3"
-id5,foo2
