--- examples/file_iterator/file_iter_test.jsonl
@@ -1,4 +1,0 @@
-{"id_str": "id1", "content": "hello"}
-{"id_str": "id3", "content": "world"}
-{"id_str": "id6", "content": "foo"}
-{"id_str": "id7", "content": "bar"}
