--- examples/file_iterator/file_iter_test.txt
@@ -1,4 +1,0 @@
-hello
-world
-foo
-bar
