--- examples/museum_gm/ground_truth.jsonl
@@ -1,9 +1,0 @@
-{"id": ["http://vocab.getty.edu/ulan/500009086", "http://data.americanartcollaborative.org/GM/constituent/charles_marion_russell"], "label": 1.0}
-{"id": ["http://vocab.getty.edu/ulan/500009086", "http://data.americanartcollaborative.org/GM/constituent/henry_joseph_sharp"], "label": 0.0}
-{"id": ["http://vocab.getty.edu/ulan/500009086", "http://data.americanartcollaborative.org/GM/constituent/kathryn_leighton_woodman"], "label": 0.0}
-{"id": ["http://vocab.getty.edu/ulan/500019643", "http://data.americanartcollaborative.org/GM/constituent/charles_marion_russell"], "label": 0.0}
-{"id": ["http://vocab.getty.edu/ulan/500019643", "http://data.americanartcollaborative.org/GM/constituent/henry_joseph_sharp"], "label": 1.0}
-{"id": ["http://vocab.getty.edu/ulan/500019643", "http://data.americanartcollaborative.org/GM/constituent/kathryn_leighton_woodman"], "label": 0.0}
-{"id": ["http://vocab.getty.edu/ulan/500012516", "http://data.americanartcollaborative.org/GM/constituent/charles_marion_russell"], "label": 0.0}
-{"id": ["http://vocab.getty.edu/ulan/500012516", "http://data.americanartcollaborative.org/GM/constituent/henry_joseph_sharp"], "label": 0.0}
-{"id": ["http://vocab.getty.edu/ulan/500012516", "http://data.americanartcollaborative.org/GM/constituent/kathryn_leighton_woodman"], "label": 1.0}
