--- examples/puam/feature_config1.json
@@ -1,11 +1,0 @@
-{
-  "missing_value_default": 0,
-  "error_handling": "ignore",
-  "features": [
-    {
-      "function": "jaro_winkler_similarity",
-      "json_path": ["name.value", "name.value"],
-      "get_first": true
-    }
-  ]
-}
