--- examples/puam/feature_config2.json
@@ -1,18 +1,0 @@
-{
-  "missing_value_default": 0,
-  "error_handling": "ignore",
-  "features": [
-    {
-      "function": "hamming_similarity",
-      "json_path": ["byear.value", "byear.value"],
-      "get_first": true
-    },
-    {
-      "function": "hybrid_jaccard_similarity",
-      "other_parameters": {"function": "soundex_similarity"},
-      "json_path": ["byear.value", "byear.value"],
-      "get_first": true,
-      "use_tokenizer": true
-    }
-  ]
-}
