--- examples/puam/feature_config3.json
@@ -1,12 +1,0 @@
-{
-  "missing_value_default": 0,
-  "error_handling": "ignore",
-  "features": [
-    {
-      "function": "hybrid_jaccard_similarity",
-      "json_path": ["name.value", "name.value"],
-      "get_first": true,
-      "use_tokenizer": true
-    }
-  ]
-}
