--- examples/puam/feature_config4.json
@@ -1,29 +1,0 @@
-{
-  "missing_value_default": 0,
-  "error_handling": "ignore",
-  "features": [
-    {
-      "function": "string_equal",
-      "json_path": ["byear.value", "byear.value"],
-      "get_first": true
-    },
-    {
-      "function": "hybrid_jaccard_similarity",
-      "other_parameters": {
-        "parameters": {"threshold": 0.5}
-      },
-      "json_path": ["name.value", "name.value"],
-      "get_first": true,
-      "use_tokenizer": true
-    },
-    {
-      "function": "hybrid_jaccard_similarity",
-      "other_parameters": {
-        "function": "soundex_similarity"
-      },
-      "json_path": ["name.value", "name.value"],
-      "get_first": true,
-      "use_tokenizer": true
-    }
-  ]
-}
