+++ rltk/io/__init__.py
@@ -1,5 +1,2 @@
-from rltk.io.reader import Reader
-from rltk.io.csv_reader import CSVReader
-from rltk.io.jsonlines_reader import JsonLinesReader
-from rltk.io.adapter import Adapter
-from rltk.io.memory_adapter import MemoryAdapter
+from rltk.io.reader import *
+from rltk.io.adapter import *
